qcacld-2.0: set rx_pending for any pending rx buffer in CE
While processing rx buffer, we force break from the per
copy engine loop once we get MAX_NUM_OF_RECEIVES per CE.
We set rx_pending to true in the above scenario, so that
rx buffer can be processed in the same tasklet context.
But rx_pending is not set to true when we have pending
buffer to process in the rx ring, when more_comp_cnt
is greater than CE_TXRX_COMP_CHECK_THRESHOLD. If there
is pending buffer in the rx ring and we hit the above
condition, pending buffer will only be processed with next
interrupt. This will defer the processing of rx packet in
the rx ring.
To avoid the above scenario, set rx_pending to true when we
have pending rx buffer in CE and sw index and read index are
not much apart.

Change-Id: I08532dd5accf8172660a84e59985a6cc610eecf9
CRs-Fixed: 1024743