qcacld-2.0: Disconnect is dropped due to invalid csr roam state
pronto to qcacld-2.0 propagation

The csr roam state is moved to joined after connection but on
receiving disconnect the state is changed to idle state even
before the disconnect indication is sent to SME. This lead to
dropping of the disconnect indication in SME as it is not
processed in idle state.

Fixed it by moving csr roam state to idle state after disconnect
done indication.

Change-Id: Ic31a2d8a0d68eaf0f4d8c6456344c78d4635f161
CRs-Fixed: 1058611