qcacld-2.0: optimize the usage of runtime PM spinlock
Analyze the critical section of the runtime PM feature
and replace the spin_lock_irqsave with spin_lock_bh.

Disabling the IRQ and usage of spinlock for runtime PM
debugging, logging and warning feature Is not required,
remove the usage of spinlock from this context.

The runtime PM APIs which is used for the prevent/allow
runtime PM suspend /resume can be use spin_lock_bh instead
of spin_lock_irqsave. For the runtime PM APIs add check point
to identify the IRQ context, If the runtime PM APIs calling
from IRQ  context report WARN_ON for debugging purpose.

CRs-Fixed: 1112447
Change-Id: If06970c6ac610a0128344e8c25a9cdd57b2542a3