qcacld-2.0: Performance optimization for HTT_T2H_MSG_TX_COMPL
When HTT_T2H_MSG_TYPE_TX_COMPL_IND is used for performance
sensitive system, the latency is harmed if ol_tx_sched is
called too late. For such platforms, TX scheduler must be
envoked as soon as possible.

Please note that this change is only for HL WLAN

Change-Id: If06a0cf5d40c59fe75bd19303a087cbf8940bae7
CRs-Fixed: 2050195